flask/01-api-auth/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
bcrypt==4.0.1
PyJWT==2.8.0
marshmallow==3.20.1
python-dotenv==1.0.0